@@ -1,33 +1,41 @@
 {
   "name": "json-rpc-engine",
   "version": "5.4.0",
-  "description": "a tool for processing JSON RPC",
+  "description": "A tool for processing JSON-RPC messages.",
   "license": "ISC",
   "author": "kumavis",
-  "main": "src/index.js",
-  "types": "src/index.d.ts",
+  "main": "dist/index.js",
+  "engines": {
+    "node": ">=10.0.0"
+  },
   "files": [
-    "src"
+    "dist"
   ],
   "scripts": {
-    "lint": "eslint . --ext js,json",
-    "lint:fix": "eslint . --ext js,json --fix",
+    "build": "tsc --project .",
+    "lint": "eslint . --ext ts,js,json",
+    "lint:fix": "eslint . --ext ts,js,json --fix",
     "test": "mocha ./test",
     "coverage": "nyc --check-coverage yarn test"
   },
   "dependencies": {
-    "eth-rpc-errors": "^3.0.0",
-    "safe-event-emitter": "^1.0.1"
+    "@metamask/safe-event-emitter": "^2.0.0",
+    "eth-rpc-errors": "^4.0.2"
   },
   "devDependencies": {
-    "@metamask/eslint-config": "^2.1.0",
-    "eslint": "^6.8.0",
+    "@metamask/eslint-config": "^4.1.0",
+    "@types/node": "^14.14.7",
+    "@typescript-eslint/eslint-plugin": "^4.8.0",
+    "@typescript-eslint/parser": "^4.8.0",
+    "eslint": "^7.13.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-json": "^2.1.0",
     "eslint-plugin-mocha": "^6.3.0",
+    "eslint-plugin-node": "^11.1.0",
     "mocha": "^7.1.1",
-    "nyc": "^15.0.0",
-    "sinon": "^9.0.2"
+    "nyc": "^15.1.0",
+    "sinon": "^9.0.2",
+    "typescript": "^4.0.5"
   },
   "repository": {
     "type": "git",
@@ -41,6 +49,7 @@
     "test": "test"
   },
   "contributors": [
-    "kumavis <aaron@kumavis.me>"
+    "kumavis <aaron@kumavis.me>",
+    "Erik Marks <rekmarks@protonmail.com>"
   ]
 }
