@@ -3,4 +3,4 @@ module.exports = {
   lines: 100,
   functions: 100,
   statements: 100,
-}
+};
